# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ChatAnswer": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "conclusion_evidence_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conclusion Evidence Id"
          },
          "evidence_ids": {
            "default": [],
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "pending": {
            "title": "Pending",
            "type": "boolean"
          },
          "plan_evidence_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Evidence Id"
          },
          "policy_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Decision"
          },
          "review_evidence_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Evidence Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "status",
          "pending"
        ],
        "title": "ChatAnswer",
        "type": "object"
      },
      "ChatInput": {
        "additionalProperties": false,
        "properties": {
          "end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "message": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "mode": {
            "default": "question",
            "enum": [
              "question",
              "task"
            ],
            "title": "Mode",
            "type": "string"
          },
          "previous_task_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Task Id"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "service_name": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,127}$",
            "title": "Service Name",
            "type": "string"
          },
          "start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          }
        },
        "required": [
          "request_id",
          "service_name",
          "message"
        ],
        "title": "ChatInput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "format": "password",
            "maxLength": 256,
            "minLength": 1,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          },
          "username": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "SessionResponse": {
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "csrf_token": {
            "title": "Csrf Token",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          }
        },
        "required": [
          "actor",
          "expires_at",
          "csrf_token"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "in": "cookie",
        "name": "ops_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Weipai AI Ops Brain",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "login_api_auth_login_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Ops-Login",
            "required": true,
            "schema": {
              "const": "1",
              "title": "X-Ops-Login",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "登录与会话"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "logout_api_auth_logout_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "title": "X-Csrf-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "登录与会话"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "me_api_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Me",
        "tags": [
          "登录与会话"
        ]
      }
    },
...
    "/api/chat": {
      "post": {
        "operationId": "chat_api_chat_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "title": "X-Csrf-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "task/evidence/delta/done 或 error 事件"
          },
          "401": {
            "description": "请先登录"
          },
          "404": {
            "description": "上一轮对话不存在"
          },
          "409": {
            "description": "相同请求 ID 的内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "任务接入暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Chat",
        "tags": [
          "AI 对话"
        ]
      }
    },
    "/api/chat/{task_id}": {
      "get": {
        "operationId": "chat_answer_api_chat__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatAnswer"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "404": {
            "description": "对话记录不存在"
          },
          "409": {
            "description": "对话输入或证据身份冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "存储或 Temporal 暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Chat Answer",
        "tags": [
          "AI 对话"
        ]
      }
    },
...
  }
}